% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bay_to_metadata.R
\name{add_bay_to_metadata}
\alias{add_bay_to_metadata}
\title{add_bay_to_metadata}
\usage{
add_bay_to_metadata(input_plate_metadata, bays)
}
\arguments{
\item{input_plate_metadata, bays}{Test test test}
}
\value{
input_plate_metadata
}
\description{
Add bay to metadata - might merge this into another function.
}
