% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Low_O2.R
\name{calculate_Low_O2}
\alias{calculate_Low_O2}
\title{calculate_Low_O2}
\usage{
calculate_Low_O2(Input.plate.data, Input.plate.metadata)
}
\arguments{
\item{Input.plate.data, Input.plate.metadata}{test test}
}
\value{
output_plate_data
}
\description{
Function that finds 5\% O2 point and adds to DF.  Needs to be 5\% of total. Might be depreciated...
}
