% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_respiration_rates.R
\name{calculate_respiration_rates}
\alias{calculate_respiration_rates}
\title{calculate_respiration_rates}
\usage{
calculate_respiration_rates(combined_plate_data, R, O2_fraction)
}
\value{
combined_plate_data
}
\description{
Function to calculate respiration rates
}
