% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_slopes.R
\name{calculate_slopes}
\alias{calculate_slopes}
\title{calculate_slopes}
\usage{
calculate_slopes(Input.plate.data, Input_plate_metadata)
}
\arguments{
\item{Input.plate.data}{test}
}
\value{
output_plate_data
}
\description{
Wrapper around Q2_Calculate_slopes - that allows runs Q2_Calculate_slopes on a loop over plates.
}
