% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_slopes_plate_loop.R
\name{calculate_slopes_plate_loop}
\alias{calculate_slopes_plate_loop}
\title{calculate_slopes_plate_loop}
\usage{
calculate_slopes_plate_loop(
  Input.plate.data,
  Input.plate.metadata,
  time_windows
)
}
\arguments{
\item{raw.data.clean, metadata, time_windows}{TEST}
}
\value{
output_plate_data
}
\description{
Fixed  030225 - Issue with well matching
}
