% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_slopes.R
\name{correct_slopes}
\alias{correct_slopes}
\title{correct_slopes}
\usage{
correct_slopes(output_plate_data, air_blank_positions, nitrogen_positions)
}
\value{
combined_plate_data
}
\description{
Function to perform slope corrections based on blanks
}
