% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_slopes_for_periods.R
\name{correct_slopes_for_periods}
\alias{correct_slopes_for_periods}
\title{correct_slopes_for_periods}
\usage{
correct_slopes_for_periods(
  output_plate_data,
  air_blank_positions,
  nitrogen_positions
)
}
\arguments{
\item{output_plate_data, air_blank_positions, nitrogen_positions}{Define overlapping time windows}
}
\value{
final_corrected_data
}
\description{
New version that calculate_slopes_for_periods
}
