% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_combined_phenotypes_df.R
\name{create_combined_phenotypes_df}
\alias{create_combined_phenotypes_df}
\title{Create Combined Phenotypes Metadata}
\usage{
create_combined_phenotypes_df()
}
\value{
A combined phenotypes data frame with unique entries.
}
\description{
This function creates default metadata for Q2 data and combines the default phenotypes data frame
with an output phenotypes data frame.
}
