% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plate_metadata.R
\name{create_plate_metadata}
\alias{create_plate_metadata}
\title{create_plate_metadata}
\usage{
create_plate_metadata(
  input_sample_metadata,
  input_plate_data,
  universal_settings,
  expected_positions
)
}
\arguments{
\item{input_sample_metadata}{test}
}
\value{
input_plate_metadata
}
\description{
Creates metadata for each plate - adds standards to each plate.
}
