% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormattedExcel.R
\name{createFormattedExcel}
\alias{createFormattedExcel}
\title{createFormattedExcel}
\usage{
createFormattedExcel(
  metadata_values = default_metadata_values,
  file_name = "Formatted_Metadata_and_Phenotypes.xlsx",
  sample_design_data = NULL,
  raw_data = NULL
)
}
\arguments{
\item{metadata_values, file_name, sample_design_data, raw_data}{Test test test}
}
\value{
Exported Excel sheet with raw/output data and metadata
}
\description{
Q2 Excel Metadata creation
}
