% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_positions.R
\docType{data}
\name{expected_positions}
\alias{expected_positions}
\title{Expected Positions for Q2 Plate Wells}
\format{
A character vector.
}
\usage{
expected_positions
}
\description{
Default metadata for Q2 Data
}
\details{
A vector of expected columns in raw/output Q2 data.
}
\keyword{datasets}
