% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_plate_metadata.R
\name{Extract_plate_metadata}
\alias{Extract_plate_metadata}
\title{Extract_plate_metadata}
\usage{
Extract_plate_metadata(
  Q2_input_metadata,
  day_col,
  plate_col,
  temp_col,
  location_col
)
}
\arguments{
\item{Q2_input_metadata, day_col, plate_col, temp_col, location_col}{Test test test}
}
\value{
df_list
}
\description{
Extract_plate_metadata
}
