% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_time_windows.R
\name{generate_time_windows}
\alias{generate_time_windows}
\title{generate_time_windows}
\usage{
generate_time_windows(start, end, step = 0.5, window_size = 2)
}
\arguments{
\item{start, end}{Define overlapping time windows}
}
\value{
time_windows
}
\description{
Define overlapping time windows
}
