% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCustomMetadataValues.R
\name{generateCustomMetadataValues}
\alias{generateCustomMetadataValues}
\title{generateCustomMetadataValues}
\usage{
generateCustomMetadataValues(
  Data_Title = "e.g Loc_TOS1_23_Q2Respiration_Rawdata",
  Summary_Info = "Griffith field trial TOS1 2023",
  TubeSize = "2 mL",
  LeafArea = "1.9 cm2",
  Light_Dark = "Dark",
  Experiment_Description = NULL,
  Collected = "Collected from 5-7 September 1782. Raw data from field experiment.",
  Performed = "Performed by XXX",
  Contact = "Contact XXX",
  ProjectID = "e.g ANU-1234-0123",
  Metadata_Excel_file_description =
    "File contains a Phenotypes sheet that describes each column of data, Sample_Design sheet that gives information on samples and their field and instrument positions, and Raw_Data sheet that contains the fluorophore output from the Q2 instrument.",
  Rights = "",
  ProjectTitle = "",
  ProjectDescription = "",
  FileDescription = "",
  FileVersion = ""
)
}
\arguments{
\item{LeafArea}{Test test test}
}
\value{
default metadata values with optional user settings
}
\description{
Function to create default metadata values with optional user settings
}
