% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IngestAndMergeXLS.R
\name{IngestAndMergeXLS}
\alias{IngestAndMergeXLS}
\title{IngestAndMergeXLS}
\usage{
IngestAndMergeXLS(InputDir)
}
\arguments{
\item{InputDir}{Test test test}
}
\value{
merged_data
}
\description{
Import all .xls files in the directory, process sheets, and merge into a single data frame
}
