% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IngestAndMergeXLS_DP22.R
\name{IngestAndMergeXLS_DP22}
\alias{IngestAndMergeXLS_DP22}
\title{IngestAndMergeXLS_DP22}
\usage{
IngestAndMergeXLS_DP22(InputDir)
}
\arguments{
\item{InputDir}{test test}
}
\value{
merged_data
}
\description{
test test
}
