% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_Calculate_slopes.R
\name{Q2_Calculate_slopes}
\alias{Q2_Calculate_slopes}
\title{Q2_Calculate_slopes}
\usage{
Q2_Calculate_slopes(raw.data.clean, metadata)
}
\arguments{
\item{raw.data.clean, R, metadata}{Test test test}
}
\value{
slopes_r_squared_df
}
\description{
Calculate slopes and R-squared values
}
