% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2IngestGRDC24.R
\name{Q2IngestGRDC24}
\alias{Q2IngestGRDC24}
\title{Q2IngestGRDC24}
\usage{
Q2IngestGRDC24(InputRawData)
}
\arguments{
\item{InputRawData}{Test test test}
}
\value{
all_sheets
}
\description{
Q2IngestGRDC24
}
