% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_phenotypes_df.R
\name{combined_phenotypes_df}
\alias{combined_phenotypes_df}
\title{Combined Phenotypes Data Frame}
\format{
A character vector.
}
\description{
Default metadata for Q2 Data
}
\details{
A vector of expected columns in raw/output Q2 data.
}
