% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_phenotypes_df.R
\docType{data}
\name{combined_phenotypes_df}
\alias{combined_phenotypes_df}
\title{Default metadata for Q2 Data}
\format{
A character vector.
\describe{
\item{combined_phenotypes_df}{Expected metadata}
combined_phenotypes_df
}
\usage{
combined_phenotypes_df
}
\description{
A vector of expected columns in raw/output Q2 data
}
\keyword{datasets}
