% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_positions.R
\docType{data}
\name{expected_positions}
\alias{expected_positions}
\title{Expected Positions for Q2 Plate Wells}
\format{
A character vector with 48 elements.
\describe{
\item{expected_positions}{Well positions in the format of rows A to F and columns 1 to 8.}
expected_positions
}
\usage{
expected_positions
}
\description{
A vector of expected positions for plate wells. These are standard positions used for well identification in Q2 runs.
}
\keyword{datasets}
